requirements.txt
opencv-python
numpy
easyocr
pillow
streamlit
